+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0 (unreleased)
 
 - Added `Transfer` model for transfers between accounts.
+- Renamed `RecurringTransaction.rrule` to `RecurringTransaction.schedule`.
 
 ## v0.1.0 (2024-12-06)